created Dashboard and Auth components Switch
