+++ client/src/components/Container/index.js
@@ -2,7 +2,17 @@ import React from 'react';
 import './style.css';
 
 function Container(props) {
-    return <div className="">{props.children}</div>;
+    return (
+    <div className="main-container">
+        <div className="row">
+            <div className="columns medium-1 centering">&nbsp;</div>
+    
+            <div className="columns medium-10 centering">{props.children}</div>
+    
+            <div className="columns medium-1 centering">&nbsp;</div>
+        </div>
+    </div>
+    )
 
 }
