+++ client/src/components/Signup/index.js
@@ -1,42 +1,44 @@
 import React from 'react';
 import './style.css';
-// import { Switch } from 'react-foundation';
 
+class Login extends React.Component {
 
+    handleLoginSuccess = (userId) => {
+        this.props.userLoggedIn(userId);
+    } 
+    render() {
+        return (
 
-class Signup extends React.Component {  
-      
-render () {  
-
-  return ( 
-    <div>
+          <div>
       <form className="callout text-center">
-        <h2>sign up</h2>
+        <h3>sign up</h3>
         <div className="floated-label-wrapper">
           <label for="full-name">full name</label>
-          <input type="text" id="full-name signupName" name="full name input" placeholder="Full name"/>
+          <input type="text" id="full-name signupName" name="full name input" placeholder="full name"/>
         </div>
         <div className="floated-label-wrapper">
           <label for="email">email</label>
-          <input type="email" id="email signupEmail" name="email input" placeholder="Email"/>
+          <input type="email" id="email signupEmail" name="email input" placeholder="email"/>
         </div>
         <div className="floated-label-wrapper">
           <label for="pass">password</label>
-          <input type="password" className="signupPass" id="pass" name="password input" placeholder="Password"/>
+          <input type="password" className="signupPass" id="pass" name="password input" placeholder="password"/>
         </div>
         <div className="floated-label-wrapper">
           <label for="pass">confirm password</label>
-          <input type="password" className="signupPass" id="pass" name="password input" placeholder="Password"/>
+          <input type="password" className="signupPass" id="pass" name="password input" placeholder="confirm password"/>
         </div>
         
         <input className="button expanded signupSubmit" type="submit" value="Sign up"/>
-        <span>already have an account? <a href="/login">Login</a> here</span><br></br>
+        <span>already have an account? &nbsp;</span> <button className="button secondary small" onClick={this.props.toggleAuthForm}>Login here</button><br></br>
       </form>
+
+      
     </div>
-
-
-    )
-  }
+           
+        )
+    }
 }
 
-export default Signup;
+export default Login;
+
